Fix maximum, minimum for `Integer` (#4335)
* fix int/long max, min value (removing decimal)

* fix integer min/max in parameter

+++ modules/swagger-codegen/src/main/java/io/swagger/codegen/languages/FlaskConnexionCodegen.java
@@ -553,20 +553,20 @@ public class FlaskConnexionCodegen extends DefaultCodegen implements CodegenConf
             example = "'" + escapeText(example) + "'";
         } else if ("Integer".equals(type) || "int".equals(type)) {
             if(p.minimum != null) {
-                example = "" + (p.minimum.intValue() + 1);
+                example = "" + (Integer.valueOf(p.minimum) + 1);
             }
             if(p.maximum != null) {
-                example = "" + p.maximum.intValue();
+                example = "" + p.maximum;
             } else if (example == null) {
                 example = "56";
             }
 
         } else if ("Long".equalsIgnoreCase(type)) {
             if(p.minimum != null) {
-                example = "" + (p.minimum.longValue() + 1);
+                example = "" + (Long.valueOf(p.minimum) + 1);
             }
             if(p.maximum != null) {
-                example = "" + p.maximum.longValue();
+                example = "" + p.maximum;
             } else if (example == null) {
                 example = "789";
             }
